+++ src/main/java/webserver/request/HttpMethod.java
@@ -1,4 +1,4 @@
-package webserver;
+package webserver.request;
 
 public enum HttpMethod {
     GET,
